Add Milvus SDK configuration, environment variables, and initial tests for user and role management
