+++ .env.testing
@@ -1,0 +1,3 @@
+MILVUS_TOKEN=root:Milvus
+MILVUS_HOST=127.0.0.1
+MILVUS_PORT=19530
