+++ tests/LaravelTest/UserTest.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Voyanara\MilvusSdk\Tests\LaravelTest;
+
+use Voyanara\MilvusSdk\Facades\Milvus;
+use Voyanara\MilvusSdk\Tests\OrchestraTestCase;
+
+class UserTest extends OrchestraTestCase
+{
+    public function test_list_users(): void
+    {
+
+    }
+}
