--- tests/Unit/CollectionTest.php
@@ -1,110 +1,0 @@
-<?php
-
-namespace Voyanara\MilvusSdk\Tests\Unit;
-
-use Voyanara\MilvusSdk\Milvus;
-use Voyanara\MilvusSdk\Tests\TestCase;
-
-class CollectionTest extends TestCase
-{
-    private Milvus $milvus;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-        $this->milvus = new Milvus('root:Milvus', 'http://localhost', '19530');
-    }
-
-    public function test_create_collection(): void
-    {
-        $schema = [
-            'fields' => [
-                [
-                    'fieldName' => 'id',
-                    'dataType' => 'Int64',
-                    'isPrimary' => true
-                ],
-                [
-                    'fieldName' => 'vector',
-                    'dataType' => 'FloatVector',
-                    'elementTypeParams' => [
-                        'dim' => '128'
-                    ]
-                ]
-            ]
-        ];
-        
-        $response = $this->milvus->collection()->createCollection('MyFirstCollection', $schema);
-        var_dump($response->body());
-        $this->assertIsArray($response->json());
-    }
-
-    public function test_drop_collection(): void
-    {
-        $collectionName = 'test_drop_collection_' . time();
-        
-        $schema = [
-            'fields' => [
-                [
-                    'fieldName' => 'id',
-                    'dataType' => 'Int64',
-                    'isPrimary' => true
-                ],
-                [
-                    'fieldName' => 'vector',
-                    'dataType' => 'FloatVector',
-                    'elementTypeParams' => [
-                        'dim' => '128'
-                    ]
-                ]
-            ]
-        ];
-        
-        $this->milvus->collection()->createCollection($collectionName, $schema);
-        
-        $response = $this->milvus->collection()->dropCollection($collectionName);
-        $this->assertIsArray($response->json());
-        $this->assertEquals(0, $response->json('code'));
-    }
-
-    public function test_add_field(): void
-    {
-        $collectionName = 'test_add_field_' . time();
-        
-        $schema = [
-            'fields' => [
-                [
-                    'fieldName' => 'id',
-                    'dataType' => 'Int64',
-                    'isPrimary' => true
-                ],
-                [
-                    'fieldName' => 'vector',
-                    'dataType' => 'FloatVector',
-                    'elementTypeParams' => [
-                        'dim' => '128'
-                    ]
-                ]
-            ]
-        ];
-        
-        $this->milvus->collection()->createCollection($collectionName, $schema);
-        
-        $fieldSchema = [
-            'fieldName' => 'new_field',
-            'dataType' => 'VarChar',
-            'nullable' => true,
-            'defaultValue' => 'default_value',
-            'elementTypeParams' => [
-                'max_length' => 255
-            ]
-        ];
-        
-        $response = $this->milvus->collection()->addField($collectionName, $fieldSchema);
-        var_dump($response->body());
-        $this->assertIsArray($response->json());
-        $this->assertEquals(0, $response->json('code'));
-        
-        $this->milvus->collection()->dropCollection($collectionName);
-    }
-}
